Removed starter code and renamed some files
