+++ server/app.js
@@ -2,17 +2,17 @@ const colors = require("colors");
 const path = require("path");
 const http = require("http");
 const express = require("express");
-const { notFound, errorHandler } = require("./middleware/error");
+const { notFound, errorHandler } = require("./middlewares/error");
 const connectDB = require("./db");
 const { join } = require("path");
 const cookieParser = require("cookie-parser");
 const logger = require("morgan");
 const passport = require("passport");
 const session = require("express-session");
-require("./utils/oauthGoogleStrategy");
+require("./utils/GoogleStrategy");
 
 const userRouter = require("./routes/user");
-const oauthRouter = require("./routes/oauth");
+const authRouter = require("./routes/auth");
 const appointmentRouter = require("./routes/appointment");
 const meetingRouter = require("./routes/meeting");
 
@@ -24,7 +24,7 @@ const server = http.createServer(app);
 
 const cookieSettings = {
   secret: process.env.SESSION_SECRET,
-  maxAge: 7 * 24 * 60 * 60 * 1000, // Session lasts for 7 days
+  maxAge: 7 * 24 * 60 * 60 * 1000 // Session lasts for 7 days
 };
 
 if (process.env.NODE_ENV === "development") {
@@ -40,7 +40,7 @@ app.use(passport.session());
 app.use(express.static(join(__dirname, "public")));
 
 // Routes
-app.use("/auth", oauthRouter);
+app.use("/auth", authRouter);
 app.use("/users", userRouter);
 app.use("/appointment", appointmentRouter);
 app.use("/meeting", meetingRouter);
