+++ server/controllers/auth.js
@@ -1,116 +1,37 @@
-const User = require("../models/User");
 const asyncHandler = require("express-async-handler");
-const generateToken = require("../utils/generateToken");
+const User = require("../models/User");
 
-// @route POST /auth/register
-// @desc Register user
+// @route POST /auth/user
+// @desc Checks if user exists
 // @access Public
-exports.registerUser = asyncHandler(async (req, res, next) => {
-  const { username, email, password } = req.body;
+exports.doesUserExist = asyncHandler(async (req, res, next) => {
+  const { email } = req.body;
 
-  const emailExists = await User.findOne({ email });
-
-  if (emailExists) {
-    res.status(400);
-    throw new Error("A user with that email already exists");
-  }
-
-  const usernameExists = await User.findOne({ username });
-
-  if (usernameExists) {
-    res.status(400);
-    throw new Error("A user with that username already exists");
-  }
-
-  const user = await User.create({
-    username,
-    email,
-    password
-  });
-
-  if (user) {
-    const token = generateToken(user._id);
-    const secondsInWeek = 604800;
-
-    res.cookie("token", token, {
-      httpOnly: true,
-      maxAge: secondsInWeek * 1000
-    });
-
-    res.status(201).json({
-      success: {
-        user: {
-          id: user._id,
-          username: user.username,
-          email: user.email
-        }
-      }
-    });
-  } else {
-    res.status(400);
-    throw new Error("Invalid user data");
-  }
-});
-
-// @route POST /auth/login
-// @desc Login user
-// @access Public
-exports.loginUser = asyncHandler(async (req, res, next) => {
-  const { email, password } = req.body;
-
-  const user = await User.findOne({ email });
-
-  if (user && (await user.matchPassword(password))) {
-    const token = generateToken(user._id);
-    const secondsInWeek = 604800;
-
-    res.cookie("token", token, {
-      httpOnly: true,
-      maxAge: secondsInWeek * 1000
-    });
-
-    res.status(200).json({
-      success: {
-        user: {
-          id: user._id,
-          username: user.username,
-          email: user.email
-        }
-      }
-    });
-  } else {
-    res.status(401);
-    throw new Error("Invalid email or password");
-  }
-});
-
-// @route GET /auth/user
-// @desc Get user data with valid token
-// @access Private
-exports.loadUser = asyncHandler(async (req, res, next) => {
-  const user = await User.findById(req.user.id);
+  const user = await User.findOne({ email: email });
 
   if (!user) {
     res.status(401);
-    throw new Error("Not authorized");
+    throw new Error(`No account exists with the email: ${email}`);
   }
 
   res.status(200).json({
     success: {
-      user: {
-        id: user._id,
-        username: user.username,
-        email: user.email
-      }
-    }
+      message: "Account exists",
+    },
   });
 });
 
-// @route GET /auth/logout
-// @desc Logout user
+// @route GET /auth/user
+// @desc Sends the logged in user's information back
 // @access Public
-exports.logoutUser = asyncHandler(async (req, res, next) => {
-  res.clearCookie("token");
+exports.loadUser = asyncHandler(async (req, res, next) => {
+  res.send(req.user);
+});
 
-  res.send("You have successfully logged out");
+// @route GET /auth/logout
+// @desc Logs the user out and redirects them to the login page
+// @access Public
+exports.logOut = asyncHandler(async (req, res, next) => {
+  req.logout();
+  res.redirect(`${process.env.CALEND_APP_DEV_URL}/login`);
 });
