--- server/middleware/isLoggedIn.js
@@ -1,5 +1,0 @@
-const isLoggedIn = (req, res, next) => {
-    req.user ? next() : res.sendStatus(401);
-}
-
-module.exports = isLoggedIn;
