+++ server/middlewares/protect.js
@@ -1,0 +1,5 @@
+const protect = (req, res, next) => {
+  req.user ? next() : res.sendStatus(401);
+};
+
+module.exports = protect;
