+++ server/routes/auth.js
@@ -1,20 +1,29 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
-const { validateRegister, validateLogin } = require('../validate');
-const protect = require('../middleware/auth');
+const passport = require("passport");
 const {
-  registerUser,
-  loginUser,
   loadUser,
-  logoutUser,
-} = require('../controllers/auth');
+  logOut,
+  loginUser,
+  doesUserExist
+} = require("../controllers/auth");
+const protect = require("../middlewares/protect");
 
-router.route('/register').post(validateRegister, registerUser);
+router.post("/user", doesUserExist);
 
-router.route('/login').post(validateLogin, loginUser);
+router.get(
+  "/google",
+  passport.authenticate("google", { scope: ["email", "profile"] }, loginUser)
+);
 
-router.route('/user').get(protect, loadUser);
+router.get(
+  "/google/callback",
+  passport.authenticate("google", {
+    successRedirect: `${process.env.CALEND_APP_DEV_URL}/dashboard`
+  })
+);
 
-router.route('/logout').get(logoutUser);
+router.get("/user", protect, loadUser);
+router.get("/logout", protect, logOut);
 
 module.exports = router;
