+++ server/routes/user.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const protect = require("../middleware/auth");
+const protect = require("../middlewares/protect");
 const { searchUsers, getUser, updateUser } = require("../controllers/user");
 
 router.route("/").get(protect, searchUsers);
