--- server/utils/generateToken.js
@@ -1,14 +1,0 @@
-const jwt = require("jsonwebtoken");
-
-const generateToken = id =>
-  jwt.sign(
-    {
-      id
-    },
-    process.env.JWT_SECRET,
-    {
-      expiresIn: "7d"
-    }
-  );
-
-module.exports = generateToken;
